---------------------------------------------------------------------- Committing in .  Modified Files:   include/fastadb.h include/hit_list.h include/misclib.h   include/partition.h include/smatrix.h lib/FSindex.c   lib/Makefile lib/bioseq.c lib/fastadb.c lib/hit_list.c   lib/keyword.c lib/pmatrix.c lib/smatrix.c   lib/cephes-double/makefile lib/cephes-double/mconf.h   progs/FSindex_create.c progs/FSindex_rand_perf.c   progs/FSindex_stats.c progs/Makefile progs/fastadb_test.c   swig/FS.i
Changes in Canada and Wellington. Too many small changes but the
algorithm stayed the same.

----------------------------------------------------------------------

+++ FSindex/src/lib/keyword.c
@@ -22,7 +22,7 @@ KW_INDEX *KW_INDEX_load_txt(const char *basename)
   KW_INDEX *KWI = callocec(1, sizeof(KW_INDEX));
 
   char *linebuf = NULL;
-  int bufsize = 0;
+  size_t bufsize = 0;
   char *bptr;
   int len;
